Added minDim
Added preference minDim and then added that preference to the methods setLevel and startLevelChange so that these  are the options that are used when fading the level of the dimmer or when using a repeat adjust level command from Rule Machine.  The default is 0 which would have the same functionality as it does today.